update naming convention
