+++ lib/screens/EntryPage/EntryPage.dart
@@ -1,17 +1,17 @@
-import 'package:flutter/material.dart';
-
 import 'package:day2day/screens/RegisterPage/RegisterPage.dart';
 import 'package:day2day/screens/EntryPage/Introduction.dart';
 import 'package:day2day/screens/LoginPage/LoginPage.dart';
 
+import 'package:flutter/material.dart';
+
 class EntryPage extends StatefulWidget {
   @override
   State<StatefulWidget> createState() {
-    return EntryPageState();
+    return _EntryPageState();
   }
 }
 
-class EntryPageState extends State<EntryPage> {
+class _EntryPageState extends State<EntryPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
