+++ lib/screens/NotificationPage/NotificationPage.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 
 class NotificationPage extends StatefulWidget {
   @override
-  NotificationPageState createState() => NotificationPageState();
+  _NotificationPageState createState() => _NotificationPageState();
 }
 
-class NotificationPageState extends State<NotificationPage> {
+class _NotificationPageState extends State<NotificationPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
